MLIBZ-1765: Swift 3.1 compatibility fixes
