+++ Cartfile.resolved
@@ -2,7 +2,7 @@ github "mattgallagher/CwlPreconditionTesting" "4a4bc648de0f975ab3e04614a25bf26b3
 github "kif-framework/KIF" "v3.5.1"
 github "kishikawakatsumi/KeychainAccess" "v3.0.1"
 github "tjboneman/NSPredicate-MongoDB-Adaptor" "2444d4a790527eb5c9fcb4e4f7b4af417048ae18"
-github "Hearst-DD/ObjectMapper" "2.2.3"
-github "mxcl/PromiseKit" "4.1.7"
+github "Hearst-DD/ObjectMapper" "2.2.5"
+github "mxcl/PromiseKit" "4.1.8"
 github "DaveWoodCom/XCGLogger" "Version_4.0.0"
-github "realm/realm-cocoa" "v2.4.3"
+github "realm/realm-cocoa" "v2.5.0"
